Optimisations des tests fonctionnels

+++ tests/func/test-runner.sh
@@ -25,10 +25,15 @@ export STRASS_TEST_REPORTS=${STRASS_TEST_REPORTS-.}
 # de test dans le PATH python.
 export PYTHONPATH=${PYTHONPATH:-.}:tests/func/site-packages/
 
+# Si non vide, alors on ne supprime pas les données, on saute les tests déjà
+# passés.
+REUSE=${REUSE-}
+
 # Si non vide, alors on attends une interraction utilisateur après les tests,
 # avant de tuer le serveur PHP et de nettoyer les données. Ça permet de
 # naviguer dedans et par exemple rédiger les tests, exporter les données, etc.
-BREAKPOINT=${BREAKPOINT-}
+# Par défaut, is REUSE est activé, on attends à la fin des tests.
+BREAKPOINT=${BREAKPOINT-${REUSE}}
 
 teardown() {
     exit_code=$1
@@ -36,7 +41,7 @@ teardown() {
     # Activer les breakpoint uniquement dans un shell interacti.
     if [ -t 1 -a -n "${BREAKPOINT}" ] ; then
         echo "Type ENTER to kill PHP and clean temporary files." >&2
-        read
+        read || true
     fi
 
     # Tuer PHP
@@ -45,14 +50,28 @@ teardown() {
 
     if [ "$exit_code" != '0' ] ; then
         # En cas d'erreur, afficher tout les logs.
-        sed 's,^,PHP: ,g' ${SERVER_LOG} >&2
-        sed 's,^,SELENIUM: ,g' ghostdriver.log >&2
+        test -f ${SERVER_LOG} && sed 's,^,PHP: ,g' ${SERVER_LOG} >&2
+        test -f ghostdriver.log && sed 's,^,SELENIUM: ,g' ghostdriver.log >&2
     fi
 
-    # Nettoyer systématiquement.
-    rm -f ${SERVER_LOG} ghostdriver.log
-    rm -rf ${STRASS_ROOT}
+    if [ -z "${REUSE}" ] ; then
+        rm -f ${SERVER_LOG} ghostdriver.log
+        rm -rf ${STRASS_ROOT}
+    fi
 }
+
+# Si on a installé notre propre phantomjs, l'injecter dans le PATH.
+if [ -d phantomjs/bin ] ; then
+    export PATH=$(readlink -f phantomjs/bin):$PATH
+fi
+
+if [ -n "${REUSE}" ] ; then
+    UNITTEST='strass.tests'
+else
+    UNITTEST='unittest'
+    # Supprimer l'état des tests. Car on va supprimer les données !
+    rm -rf .reuse-state ${STRASS_ROOT}
+fi
 
 UNITTEST_ARGS="--verbose"
 if [ -n "${BREAKPOINT}" ] ; then
@@ -64,8 +83,8 @@ fi
 libsqlite=$(find /usr/lib -name "*libsqlite3.so")
 libfaketime=$(dpkg -L libfaketime| grep libfaketime.so)
 
-# Démarrer le serveur le 9 août 2007 à 9h32m08. Accélérer le temps par 10.
-LD_PRELOAD="${libfaketime} ${libsqlite}" FAKETIME="@2007-08-09 09:32:08 x10" \
+# Démarrer le serveur le 9 août 2007 à 9h32m08. Accélérer le temps par 100.
+LD_PRELOAD="${libfaketime} ${libsqlite}" FAKETIME="@2007-08-09 09:32:08 x100" \
           maint/scripts/serve.sh &>${SERVER_LOG} &
 PHP_PID=$!
 echo "PHP PID is ${PHP_PID}" >&2
@@ -73,4 +92,4 @@ echo "PHP PID is ${PHP_PID}" >&2
 trap 'teardown $?' EXIT QUIT INT TERM ABRT ALRM HUP CHLD
 
 # On délègue à python3 unittest les tests. Usuel.
-python3 -m unittest discover ${UNITTEST_ARGS} tests/func/${TESTCASE}/
+python3 -m ${UNITTEST} discover ${UNITTEST_ARGS} tests/func/${TESTCASE}/
